kernel: move type-specific traverse code to better places
E.g. the code for traversing and copying weak pointer objects now is in
weakptr.c, and traverse.c knows nothing about weak pointer objects.
